backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
sqlalchemy>=2.0.30,<3.0.0
pydantic[email]>=2.7.0,<3.0.0
python-dotenv>=1.0.1,<2.0.0
psycopg[binary]>=3.2.0,<4.0.0

